+++ A-apps/the-boring-rpg/client-browser/src/services/loader.js
@@ -4,12 +4,13 @@ import get_loader from '@offirmo-private/iframe-loading'
 import { ↆcordova } from './cordova'
 import logger from './logger'
 
-ↆcordova.then(() => {
+
+ↆcordova.finally(() => {
 	// https://cordova.apache.org/docs/en/latest/reference/cordova-plugin-splashscreen/index.html#preferences
 	if ('splashscreen' in navigator) {
 		navigator.splashscreen.hide()
 	}
-})
+}).catch(() => {})
 
 
 const declare_app_loaded = memoize_one(function _declare_app_loaded() {