+++ 1-stdlib/deferred/CHANGELOG.md
@@ -2,6 +2,7 @@
 **This package follows [semver](https://semver.org/).**
 
 ## PENDING
+* [chore] [breaking] will now by default prevent the "uncaught rejection" message
 * [chore] [breaking] cjs pre-build now targeting node12/ES2019 [details](../../CONTRIBUTING/module-exports.md)
 * [feat] slightly improved typings, following the latest Promise definition (Typescript 3.8.3)
 * [doc] credits